refactor(structure): jest test on ts files without build
