+++ package.json
@@ -12,6 +12,7 @@
     "s": "npx nps",
     "s:list": "yarn s --scripts",
     "build": "yarn s build",
+    "clean": "rimraf packages/*/lib",
     "pretest": "npm run db:prepare:typeorm",
     "test": "yarn s test",
     "pretest:coveralls": "yarn pretest",
@@ -85,6 +86,7 @@
     "swagger-ui-express": "4.0.6",
     "ts-jest": "24.0.2",
     "ts-node": "8.3.0",
+    "tsconfig-extends": "^1.0.1",
     "tsconfig-paths": "3.8.0",
     "tslint": "5.18.0",
     "tslint-config-prettier": "1.18.0",
@@ -92,7 +94,6 @@
     "typescript": "3.3.3",
     "validate-commit-msg": "2.14.0"
   },
-  "devDependencies": {},
   "author": {
     "name": "Michael Yali",
     "email": "mihon4ik@gmail.com"
