+++ packages/crud-request/src/index.ts
@@ -1,3 +1,5 @@
 export * from './request-query.builder';
+export * from './request-query.parser';
 export * from './interfaces';
 export * from './types';
+export * from './exceptions';
