+++ packages/crud-typeorm/test/__fixture__/companies.service.ts
@@ -1,8 +1,8 @@
 import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 
-import { TypeOrmCrudService } from '../../../crud-typeorm/src/typeorm-crud.service';
 import { Company } from '../../../../integration/crud-typeorm/companies';
+import { TypeOrmCrudService } from '../../src/typeorm-crud.service';
 
 @Injectable()
 export class CompaniesService extends TypeOrmCrudService<Company> {
