+++ packages/crud-typeorm/test/__fixture__/projects.service.ts
@@ -1,8 +1,8 @@
 import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 
-import { TypeOrmCrudService } from '../../../crud-typeorm/src/typeorm-crud.service';
 import { Project } from '../../../../integration/crud-typeorm/projects';
+import { TypeOrmCrudService } from '../../src/typeorm-crud.service';
 
 @Injectable()
 export class ProjectsService extends TypeOrmCrudService<Project> {
