+++ packages/crud/src/index.ts
@@ -4,3 +4,4 @@ export * from './interfaces';
 export * from './types';
 export * from './module';
 export * from './interceptors';
+export * from './services';
